Create check on payload length. Currently on billservice only

+++ norrin/tests.py
@@ -1,27 +1,54 @@
 from nose.tools import nottest
+import logging
+
+import os
+
+import json
 
 from mongokit import ObjectId
 
 from norrin.notifications.models import connection
-from norrin.notifications.services import BillService, VoteService, BillActionService
+from norrin.notifications.services import BillService, VoteService, \
+    BillActionService, adapters
 
 from sunlight import congress
 from dateutil.parser import parse as parse_date
+import urbanairship as ua
 
 TEST_DATABASE = 'norrinTests'
+PRESERVE_TEST_DB = os.environ.get('PRESERVE_TEST_DB', False)
 
-def tearDownModule(self):
-    connection.drop_database(TEST_DATABASE)
+logger = logging.getLogger(__name__)
+
+
+class InspectNotificationAdapter(object):
+    def __init__(self):
+        self.payloads = []
+
+    def push(self, notification):
+        pyld_json = json.dumps(ua.notification(ios=ua.ios(alert=notification.message, extra=notification.context)))
+        self.payloads.append(pyld_json)
+
+def teardown(self):
+    if PRESERVE_TEST_DB:
+        logger.warning('Preserving test database...')
+    else:
+        logger.warning('Dropping test database...')
+        connection.drop_database(TEST_DATABASE)
+
 
 class DBConnected(object):
     """docstring for Base"""
     def __init__(self):
         self.db = connection[TEST_DATABASE]
 
+
 class TestBills(DBConnected):
 
     def setup(self):
-        BillService(self.db).load_data()
+        self.service = BillService(self.db)
+        self.service.load_data()
+
         if self.db.bills.count() == 0:
             last_bills = congress.bills(order='introduced_on', fields='bill_id,sponsor_id,introduced_on', per_page=20)
             for bill in last_bills:
@@ -66,6 +93,17 @@ class TestBills(DBConnected):
         obj.update(self.bill_update_dict())
         obj.save()
         assert obj.bill_id != False
+
+    def test_bill_notification_size(self):
+        '''Check size of notification payloads for APNS using the InspectNotificationAdapter'''
+        adapter = InspectNotificationAdapter()
+        adapters.register(adapter)
+        self.service.run()
+        logger.warning('Checking payloads')
+        for payload in adapter.payloads:
+            pyld_len = len(payload.encode('utf-8'))
+            logger.warning(u'Payload length: {}'.format(pyld_len))
+            assert pyld_len <= 256
 
 
 class TestBillActions(DBConnected):
